\name{aids}
\alias{aids}
\non_function{}
\title{data from Section 1.19}
\description{
  The \code{aids} data frame has 295 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{infect}{
      Infection time for AIDS, years
    }
    \item{induct}{
      Induction time for AIDS, years
    }
    \item{adult}{
      Indicator of adult (1=adult, 0=child)
    }
  }
}
\source{
  Klein and Moeschberger (1997) \emph{Survival Analysis Techniques for Censored
    and truncated data}, Springer.
  Lagakos et al. Biometrika 68 (1981): 515-523. 
}
\examples{
data(aids)
}
\keyword{datasets}
