\name{baboon}
\alias{baboon}
\non_function{}
\title{data from Exercise 5.8, p147}
\description{
  The \code{baboon} data frame has 25 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{date}{
      Date (day/month/year)
    }
    \item{time}{
      Descent time (military time)
    }
    \item{observed}{
      Indicator of observed or not (1=observed, 0=not observed)
    }
  }
}
\source{
  Klein and Moeschberger (1997) \emph{Survival Analysis Techniques for Censored
    and truncated data}, Springer.
}
\examples{
data(baboon)
}
\keyword{datasets}
