\name{rats}
\alias{rats}
\non_function{}
\title{data from Exercise 7.13, p225}
\description{
  The \code{rats} data frame has 50 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{time}{
      Time to tumor development
    }
    \item{tumor}{
      Indicator of tumor development (1=yes, 0=no)
    }
    \item{trt}{
      Treatment (1=treated with drug, 0=given placebo)
    }
    \item{litter}{
      Litter
    }
  }
}
\source{
  Klein and Moeschberger (1997) \emph{Survival Analysis Techniques for Censored
    and truncated data}, Springer.
}
\examples{
data(rats)
}
\keyword{datasets}
