\name{stddiag}
\alias{stddiag}
\non_function{}
\title{data from Exercise 5.6, p146}
\description{
  The \code{stddiag} data frame has 25 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{encounter}{
      Months from 1/93 to encounter
    }
    \item{diagnosed}{
      Months until STD diagnosed in the clinic
    }
  }
}
\source{
  Klein and Moeschberger (1997) \emph{Survival Analysis Techniques for Censored
    and truncated data}, Springer.
}
\examples{
data(stddiag)
}
\keyword{datasets}
